refactor(0.0.8) : 将工具注册代码从server.ts拆分到独立文件
将所有MCP工具注册代码从server.ts文件抽取到新的src/tools目录下的独立模块中，按功能分类整理。这样改进了代码组织结构，提高了可维护性，并使server.ts文件更加简洁，专注于服务器配置和连接处理。
